[rubis] (bug-fix:minor) Attribute 'maxBid' is a float, not an int.

+++ src/radlab/rain/workload/rubis/model/RubisItem.java
@@ -52,7 +52,7 @@ public class RubisItem
 	public float reservePrice;
 	public float buyNow;
 	public int nbOfBids;
-	public int maxBid;
+	public float maxBid;
 	public Date startDate;
 	public Date endDate;
 	public int seller;
